safari-600.7-branch/LayoutTests/java/lc3/JavaObject/JavaObjectToBoolean-001-n-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion JavaObject to boolean

FAIL dt.PUB_BOOLEAN_REPRESENTATION = false ;dt.setBoolean( dt ); dt.PUB_BOOLEAN should throw an exception. Was false.
FAIL dt.PUB_BOOLEAN_REPRESENTATION = false ;dt.setBoolean( dt ); dt.getBoolean() should throw an exception. Was false.
PASS typeof dt.getBoolean) threw exception SyntaxError: Unexpected token ')'.
PASS successfullyParsed is true

TEST COMPLETE
 
